Django-back/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Markdown==3.3.6
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
